Version 0.1 Commit
So here is version 0.1 of sinkrontrack-server-go
It has authentication, the ability to add playlists
Create new users.
Complete with tests

Authentication is implemented with JWT
Added token refresh, and token authentication on requests that
will require it.
Added pre-seeding a userLogin account for the admin user
Creating routing
Adding check for environment var on startup

Adding a lock on playlists so that the first user has a renewable
1hr lock on updating playlist position.
Working on adding in testing, I need to move database access to
an interface, so I can implement a mock database as part of testing

Moved contents of internal/model package into internal/storage so I
can make the db more agnostic, and also prep it for writing unit tests

Adding the ability to set and update the "AdminUser" flag and checking
to see if the user is an Admin while attempting to update a user record
Moved all of the error handling to my returnError function call